+++ src/Grpc.Net.ClientFactory/InterceptorLifetime.cs
@@ -1,0 +1,37 @@
+﻿#region Copyright notice and license
+
+// Copyright 2019 The gRPC Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#endregion
+
+using Grpc.Core.Interceptors;
+
+namespace Grpc.Net.ClientFactory
+{
+    /// <summary>
+    /// Specifies the lifetime of an <see cref="Interceptor"/> for a client.
+    /// </summary>
+    public enum InterceptorLifetime
+    {
+        /// <summary>
+        /// Specifies that the <see cref="Interceptor"/> will be resolved every time a client instance is created.
+        /// </summary>
+        Client,
+        /// <summary>
+        /// Specifies that a single instance of the <see cref="Interceptor"/> will be resolved for each channel.
+        /// </summary>
+        Channel
+    }
+}
